integrate(persistence): assemble validated P1 authorities
