--- src/google_work_agent/adapters/persistence/secret_boundary.py
@@ -1,36 +1,0 @@
-"""Persistence-side defense in depth for Trace/Audit secret boundaries."""
-
-from __future__ import annotations
-
-from dataclasses import replace
-
-from google_work_agent.adapters.persistence.repositories import (
-    SQLiteAuditRepository,
-    SQLiteTraceRepository,
-)
-from google_work_agent.application.observability import sanitize_persistent_event_json
-from google_work_agent.ports import AuditEventRecord, TraceEventRecord
-
-
-class SecretBoundaryAuditRepository(SQLiteAuditRepository):
-    """Audit repository that sanitizes event JSON immediately before SQLite persistence."""
-
-    def add(self, event: AuditEventRecord) -> None:
-        super().add(
-            replace(
-                event,
-                metadata_json=sanitize_persistent_event_json(event.metadata_json),
-            )
-        )
-
-
-class SecretBoundaryTraceRepository(SQLiteTraceRepository):
-    """Trace repository that sanitizes event JSON immediately before SQLite persistence."""
-
-    def add(self, event: TraceEventRecord) -> None:
-        super().add(
-            replace(
-                event,
-                payload_json=sanitize_persistent_event_json(event.payload_json),
-            )
-        )
